+++ CHANGELOG.md
@@ -99,6 +99,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - [#623] Fixed an issue where Next triggered the "uncontrolled to controlled" warning unnecessarily. Removed a `useEffect` call from `usePrevious` and a ref that tracked "first render" from `useQueryBuilderSchema`.
 - [#625] A default value will not be selected unnecessarily when `valueEditorType` evaluates to `"multiselect"`.
 - Refactored custom hooks to avoid unnecessary `useEffect` calls.
+- [#663] `formatQuery` for "jsonlogic" will no longer collapse subgroups that only contain one rule into a single JsonLogic rule. Full query objects that contain only one rule will still be collapsed.
+- [#663] When a `values` list is defined for a field, and the value is reset due to `resetOnFieldChange` or `resetOnOperatorChange` being `true`, the rule `value` will no longer be set to the first value in the list unless the `valueEditorType` evaluates to "select" or "radio".
 
 ## [v6.5.5] - 2024-01-15
 
@@ -1529,6 +1531,7 @@ Maintenance release focused on converting to a monorepo with Vite driving the bu
 [#646]: https://github.com/react-querybuilder/react-querybuilder/pull/646
 [#653]: https://github.com/react-querybuilder/react-querybuilder/pull/653
 [#654]: https://github.com/react-querybuilder/react-querybuilder/pull/654
+[#663]: https://github.com/react-querybuilder/react-querybuilder/pull/663
 
 <!-- Release comparison links -->